Add files via upload
With a dataset of fish species, with some of it characteristic like it vertical, diagonal, length, height, and width. We will try to predict the weight of the fish based on their characteristic. We will use Linear Regression Method to see whether the weight of the fish related to their characteristic.
Linear Regression:
Linear regression is also a type of machine-learning algorithm more specifically a supervised machine-learning algorithm that learns from the labelled datasets and maps the data points to the most optimized linear functions. which can be used for prediction on new datasets.
1.Species: Species name of fish
2.Weight: Weight of fish in gram
3.Length1: Vertical length in cm
4.Length2: Diagonal length in cm
5.Length3: Cross length in cm
6.Height: Height in cm
7.Width: Diagonal width in cm